requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.5
django-ckeditor==6.5.1
django-compat==1.0.15
django-coupon-management==0.3.2
django-coupons==1.1.0
django-extra-views==0.14.0
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-render-partial==0.4
django-robots==5.0
django-throttle-requests==0.6.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
idna==3.4
ippanel==2.0.7
jdatetime==4.1.0
kombu==5.2.4
Pillow==9.4.0
prompt-toolkit==3.0.36
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
validate-email==1.3
vine==5.0.0
wcwidth==0.2.6
